Create total_energy_extraction.py
This file takes in a parent directory and finds subfolders with same prefix and loop through them in an orderly fashion. Then it picks out the log.lammps file and extracts the values at the total energy column and save it to an array called "energy". Then it plots the energy values for all the subfolder energy values.